backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-generativeai>=0.7.0
google-genai>=0.3.0
pandas>=2.0.0
python-multipart>=0.0.9
aiofiles>=23.2.1
python-dotenv>=1.0.0
